+++ packages/cadl-ranch-api/src/expectation.ts
@@ -7,6 +7,7 @@ import {
   validateCoercedDateBodyEquals,
   validateHeader,
   validateQueryParam,
+  validateXmlBodyEquals,
 } from "./request-validations.js";
 import { CollectionFormat, RequestExt } from "./types.js";
 import { ValidationError } from "./validation-error.js";
@@ -91,5 +92,15 @@ export class RequestExpectation {
     if (!deepEqual(actual, expected, { strict: true })) {
       throw new ValidationError(message, expected, actual);
     }
+  }
+
+  /**
+   * Expect the body of the request to be semantically equivalent to the provided XML string.
+   * The XML declaration prefix will automatically be added to expectedBody.
+   * @param expectedBody expected value of request body.
+   * @throws {ValidationError} if there is an error.
+   */
+  public xmlBodyEquals(expectedBody: string): void {
+    validateXmlBodyEquals(this.originalRequest, expectedBody);
   }
 }