+++ packages/cadl-ranch-api/src/response-utils.ts
@@ -11,3 +11,16 @@ export function json(content: unknown): MockResponseBody {
     rawContent: JSON.stringify(content),
   };
 }
+
+/**
+ * Sends the provided XML string in a MockResponse body.
+ * The XML declaration prefix will automatically be added to xmlString.
+ * @content Object to return as XML.
+ * @returns {MockResponseBody} response body with application/xml content type.
+ */
+export function xml(xmlString: string): MockResponseBody {
+  return {
+    contentType: "application/xml",
+    rawContent: `<?xml version='1.0' encoding='UTF-8'?>` + xmlString,
+  };
+}