+++ packages/cadl-ranch-specs/http/parameters/spread/mockapi.ts
@@ -10,6 +10,36 @@ Scenarios.Parameters_Spread_Model_spreadAsRequestBody = passOnSuccess(
   }),
 );
 
+Scenarios.Parameters_Spread_Model_spreadCompositeRequestOnlyWithBody = passOnSuccess(
+  mockapi.put("/parameters/spread/model/composite-request-only-with-body", (req) => {
+    req.expect.bodyEquals({ name: "foo" });
+    return { status: 204 };
+  }),
+);
+
+Scenarios.Parameters_Spread_Model_spreadCompositeRequestWithoutBody = passOnSuccess(
+  mockapi.put("/parameters/spread/model/composite-request-without-body/foo", (req) => {
+    req.expect.containsHeader("test-header", "bar");
+    return { status: 204 };
+  }),
+);
+
+Scenarios.Parameters_Spread_Model_spreadCompositeRequest = passOnSuccess(
+  mockapi.put("/parameters/spread/model/composite-request/foo", (req) => {
+    req.expect.containsHeader("test-header", "bar");
+    req.expect.bodyEquals({ name: "foo" });
+    return { status: 204 };
+  }),
+);
+
+Scenarios.Parameters_Spread_Model_spreadCompositeRequestMix = passOnSuccess(
+  mockapi.put("/parameters/spread/model/composite-request-mix/foo", (req) => {
+    req.expect.containsHeader("test-header", "bar");
+    req.expect.bodyEquals({ prop: "foo" });
+    return { status: 204 };
+  }),
+);
+
 Scenarios.Parameters_Spread_Alias_spreadAsRequestBody = passOnSuccess(
   mockapi.put("/parameters/spread/alias/request-body", (req) => {
     req.expect.bodyEquals({ name: "foo" });
@@ -29,13 +59,27 @@ Scenarios.Parameters_Spread_Alias_spreadWithMultipleParameters = passOnSuccess(
   mockapi.put("/parameters/spread/alias/multiple-parameters/1", (req) => {
     req.expect.containsHeader("x-ms-test-header", "bar");
     req.expect.bodyEquals({
-      prop1: "foo1",
-      prop2: "foo2",
-      prop3: "foo3",
-      prop4: "foo4",
-      prop5: "foo5",
-      prop6: "foo6",
+      requiredString: "foo",
+      optionalInt: 1,
+      requiredIntList: [1, 2],
+      optionalStringList: ["foo", "bar"],
     });
+    return { status: 204 };
+  }),
+);
+
+Scenarios.Parameters_Spread_Alias_spreadParameterWithInnerModel = passOnSuccess(
+  mockapi.post("/parameters/spread/alias/inner-model-parameter/1", (req) => {
+    req.expect.containsHeader("x-ms-test-header", "bar");
+    req.expect.bodyEquals({ name: "foo" });
+    return { status: 204 };
+  }),
+);
+
+Scenarios.Parameters_Spread_Alias_spreadParameterWithInnerAlias = passOnSuccess(
+  mockapi.post("/parameters/spread/alias/inner-alias-parameter/1", (req) => {
+    req.expect.containsHeader("x-ms-test-header", "bar");
+    req.expect.bodyEquals({ name: "foo", age: 1 });
     return { status: 204 };
   }),
 );