+++ packages/cadl-ranch-specs/http/type/property/optionality/mockapi.ts
@@ -77,6 +77,18 @@ Scenarios.Type_Property_Optional_Duration_getDefault = passOnSuccess(durationMoc
 Scenarios.Type_Property_Optional_Duration_putAll = passOnSuccess(durationMock.putAll);
 Scenarios.Type_Property_Optional_Duration_putDefault = passOnSuccess(durationMock.putDefault);
 
+const plainDateMock = createMockApis("plainDate", "2022-12-12");
+Scenarios.Type_Property_Optional_PlainDate_getAll = passOnSuccess(plainDateMock.getAll);
+Scenarios.Type_Property_Optional_PlainDate_getDefault = passOnSuccess(plainDateMock.getDefault);
+Scenarios.Type_Property_Optional_PlainDate_putAll = passOnSuccess(plainDateMock.putAll);
+Scenarios.Type_Property_Optional_PlainDate_putDefault = passOnSuccess(plainDateMock.putDefault);
+
+const plainTimeMock = createMockApis("plainTime", "13:06:12");
+Scenarios.Type_Property_Optional_PlainTime_getAll = passOnSuccess(plainTimeMock.getAll);
+Scenarios.Type_Property_Optional_PlainTime_getDefault = passOnSuccess(plainTimeMock.getDefault);
+Scenarios.Type_Property_Optional_PlainTime_putAll = passOnSuccess(plainTimeMock.putAll);
+Scenarios.Type_Property_Optional_PlainTime_putDefault = passOnSuccess(plainTimeMock.putDefault);
+
 const collectionsBytesMock = createMockApis("collections/bytes", ["aGVsbG8sIHdvcmxkIQ==", "aGVsbG8sIHdvcmxkIQ=="]);
 Scenarios.Type_Property_Optional_CollectionsByte_getAll = passOnSuccess(collectionsBytesMock.getAll);
 Scenarios.Type_Property_Optional_CollectionsByte_getDefault = passOnSuccess(collectionsBytesMock.getDefault);
@@ -101,7 +113,7 @@ Scenarios.Type_Property_Optional_IntLiteral_getDefault = passOnSuccess(intLitera
 Scenarios.Type_Property_Optional_IntLiteral_putAll = passOnSuccess(intLiteralMock.putAll);
 Scenarios.Type_Property_Optional_IntLiteral_putDefault = passOnSuccess(intLiteralMock.putDefault);
 
-const floatLiteralMock = createMockApis("float/literal", 1.2);
+const floatLiteralMock = createMockApis("float/literal", 1.25);
 Scenarios.Type_Property_Optional_FloatLiteral_getAll = passOnSuccess(floatLiteralMock.getAll);
 Scenarios.Type_Property_Optional_FloatLiteral_getDefault = passOnSuccess(floatLiteralMock.getDefault);
 Scenarios.Type_Property_Optional_FloatLiteral_putAll = passOnSuccess(floatLiteralMock.putAll);
@@ -125,7 +137,7 @@ Scenarios.Type_Property_Optional_UnionIntLiteral_getDefault = passOnSuccess(unio
 Scenarios.Type_Property_Optional_UnionIntLiteral_putAll = passOnSuccess(unionIntLiteralMock.putAll);
 Scenarios.Type_Property_Optional_UnionIntLiteral_putDefault = passOnSuccess(unionIntLiteralMock.putDefault);
 
-const unionFloatLiteralMock = createMockApis("union/float/literal", 2.3);
+const unionFloatLiteralMock = createMockApis("union/float/literal", 2.375);
 Scenarios.Type_Property_Optional_UnionFloatLiteral_getAll = passOnSuccess(unionFloatLiteralMock.getAll);
 Scenarios.Type_Property_Optional_UnionFloatLiteral_getDefault = passOnSuccess(unionFloatLiteralMock.getDefault);
 Scenarios.Type_Property_Optional_UnionFloatLiteral_putAll = passOnSuccess(unionFloatLiteralMock.putAll);