+++ packages/cadl-ranch-specs/http/versioning/removed/mockapi.ts
@@ -1,0 +1,14 @@
+import { passOnSuccess, mockapi, json } from "@azure-tools/cadl-ranch-api";
+import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
+
+export const Scenarios: Record<string, ScenarioMockApi> = {};
+
+Scenarios.Versioning_Removed_v2 = passOnSuccess(
+  mockapi.post("/versioning/removed/api-version:v2/v2", (req) => {
+    req.expect.bodyEquals({ prop: "foo", enumProp: "enumMemberV2", unionProp: "bar" });
+    return {
+      status: 200,
+      body: json({ prop: "foo", enumProp: "enumMemberV2", unionProp: "bar" }),
+    };
+  }),
+);