+++ packages/typespec-bump-deps/src/cli/typespec-bump-deps.ts
@@ -9,8 +9,12 @@ const knownPackages = [
   "@typespec/openapi",
   "@typespec/rest",
   "@typespec/versioning",
+  "@typespec/xml",
   "@azure-tools/typespec-client-generator-core",
   "@azure-tools/typespec-azure-core",
+  "@azure-tools/typespec-azure-resource-manager",
+  "@azure-tools/typespec-azure-rulesets",
+  "@azure-tools/typespec-autorest",
   "@typespec/eslint-config-typespec",
   "@typespec/library-linter",
 ];