+++ packages/typespec-bump-deps/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @azure-tools/typespec-bump-deps
 
+## 0.9.5
+
+### Patch Changes
+
+- df27070: Add `@typespec/xml`
+
+## 0.9.4
+
+### Patch Changes
+
+- bbd186f: Add typespec-azure-rulesets to typespec-bump-deps
+
+## 0.9.3
+
+### Patch Changes
+
+- a14450b: Include typespec-autorest in packages, it is used by typespec-azure-resource-manager
+
+## 0.9.2
+
+### Patch Changes
+
+- 7161c8c: feat(typespec): bump to 0.55.0
+
+## 0.9.1
+
+### Patch Changes
+
+- b871ba8: chore(typespec): bump to 0.54.0
+
+## 0.9.0
+
+### Minor Changes
+
+- f2a9bce: bump tsp package versions to 0.53.0
+
+## 0.8.1
+
+### Patch Changes
+
+- 68a7bde: Add typespec-azure-resource-manager package to typespec-bump-deps
+
 ## 0.8.0
 
 ### Minor Changes